separated profile/author, +profile operations

+++ Assignment2/app/views/Profile/index.php
@@ -14,56 +14,59 @@
 
     <div class="container mb-5"> 
         <h2>List of Publications</h2>
-
-        <table class="table table-striped table-bordered table-hover">
-            <tr>
-                <td>Title</td>
-                <td>Timestamp</td>
-                <td>Status</td>
-                <td>Change Status</td>
-                <td colspan="2" class="text-center">Actions</td>
-            </tr>
-
             <?php
-                foreach($data["publications"] as $publication){
-                    echo"<tr>";
+                if (!($data["publications"])) {
+                    echo "No publications found...";
+                }
+                else {
+                    echo 
+                    '<table class="table table-striped table-bordered table-hover">
+                    <tr>
+                        <td>Title</td>
+                        <td>Timestamp</td>
+                        <td>Status</td>
+                        <td>Change Status</td>
+                        <td colspan="2" class="text-center">Actions</td>
+                    </tr>';
 
-                    echo"<td><a href='/Assignment2/Publication/getPublication/$publication->publication_id'>$publication->publication_title</td>";
-                    echo"<td>$publication->timestamp</td>";
-                    echo"<td>";
-                    if ($publication->publication_status == 0) {
-                        echo "Private";
+                    foreach($data["publications"] as $publication){
+                        echo"<tr>";
+    
+                        echo"<td><a href='/Assignment2/Publication/getPublication/$publication->publication_id'>$publication->publication_title</td>";
+                        echo"<td>$publication->timestamp</td>";
+                        echo"<td>";
+                        if ($publication->publication_status == 0) {
+                            echo "Private";
+                        }
+                        else {
+                            echo "Public";
+                        }
+                        echo "</td>";
+    
+                        echo"<td>";
+                        if ($publication->publication_status == 0) {
+                            echo "<a href='/Assignment2/Publication/makePublicationPublic/$publication->publication_id'>Make Public</a>";
+                        }
+                        else {
+                            echo "<a href='/Assignment2/Publication/makePublicationPrivate/$publication->publication_id'>Make Private</a>";
+                        }
+                        echo "</td>";
+    
+                        echo
+                        "<td>
+                            <a href='/Assignment2/Publication/editPublication/$publication->publication_id'>Edit</a>
+                        </td>";
+                        echo
+                        "<td>
+                            <a href='/Assignment2/Publication/deletePublication/$publication->publication_id'>Delete</a>
+                        </td>";
+    
+                        echo"</tr>";
                     }
-                    else {
-                        echo "Public";
-                    }
-                    echo "</td>";
 
-                    echo"<td>";
-                    if ($publication->publication_status == 0) {
-                        echo "<a href='/Assignment2/Publication/makePublicationPublic/$publication->publication_id'>Make Public</a>";
-                    }
-                    else {
-                        echo "<a href='/Assignment2/Publication/makePublicationPrivate/$publication->publication_id'>Make Private</a>";
-                    }
-                    echo "</td>";
-
-                    echo
-                    "<td>
-                        <a href='/Assignment2/Publication/editPublication/$publication->publication_id'>Edit</a>
-                    </td>";
-                    echo
-                    "<td>
-                        <a href='/Assignment2/Publication/deletePublication/$publication->publication_id'>Delete</a>
-                    </td>";
-
-                    echo"</tr>";
-
+                    echo "</table>";
                 }
             ?>
-        </table>
-
-
     </div>
 
     <div class="container">
